requirements.txt
ultralytics==8.1.0
gradio
Pillow
matplotlib